Use Rail's new GlobalID.
Instead of serialing the object directly in YAML (which always has random objects added in the serializations), it instead uses Rail's new GlobalID system.

This is just a sample to show that it's easily possible to use, although I would modify closer to the handler.to_yaml serialize in base.rb.

(In fact, with GlobalID, you can change ther serializer from YAML to JSON if wanted...)